GET lists

+++ src/main/java/com/example/dslist/entities/GameList.java
@@ -14,21 +14,21 @@ public class GameList {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	
 	public GameList() {}
 
-	public GameList(Integer id, String name) {
+	public GameList(Long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -56,6 +56,7 @@ public class GameList {
 		GameList other = (GameList) obj;
 		return Objects.equals(id, other.id);
 	}
+
 	
 	
 }
